+++ components/dropzone.module.css
@@ -1,30 +1,16 @@
 .container {
-  background-color: #FF1C00;
   padding: 12px;
-  margin: 50px;
+  background-color: #FF1C00;
+  height: 100%;
 }
 
 .zone {
   background-color: #D41F17;
-  border: #AA342D 3px dashed;
-  padding: 75px;
+  border: #AA342D 4px dashed;
 }
 
-.active {
-  background-color: #D41F17;
-  filter: brightness(85%); 
-  border: #AA342D 3px dashed;
-  padding: 75px;
-}
-
-.zone p {
-  color: white;
-  font-size: 12pt;
-  font-family: 'NotoSansRegular';
-}
-
-.active p {
-  color: white;
-  font-size: 12pt;
-  font-family: 'NotoSansRegular';
+.zoneActive,
+.zone:hover {
+  filter: brightness(85%);
+  cursor: pointer;
 }
