+++ package.json
@@ -11,7 +11,9 @@
     "next": "9.4.1",
     "react": "16.13.1",
     "react-dom": "16.13.1",
-    "react-dropzone": "^11.0.1"
+    "react-dropzone": "^11.0.1",
+    "react-soundplayer": "^1.0.4",
+    "wavesurfer.js": "^3.3.3"
   },
   "devDependencies": {
     "@types/node": "^14.0.1",
