+++ pages/_app.js
@@ -1,5 +1,6 @@
 import '../styles/global.css'
 import '../styles/font.css'
+import '../styles/soundplayer.css'
 
 export default function App({ Component, pageProps }) {
   return <Component {...pageProps} />
